+++ index.js
@@ -19,6 +19,7 @@ const client = new Client({
 
 // --- Own Modules --- \\
 const wordle = require("./src/modules/wordle");
+const words = require("./src/modules/words");
 const embeds = require("./src/modules/embeds");
 
 // --- Variables --- \\
@@ -35,7 +36,37 @@ client.on('ready', _ => {
   client.user.setActivity("Wordle on Discord", {
     type: "PLAYING"
   });
+
+  // --- Register commands --- \\
+  const guild = client.guilds.cache.get("908019966022942730");
+  var commands = guild.commands; // --- Guild commands --- \\
+  // var commands = client.application.commands; // --- Global commands --- \\
+  commands.create({
+    name: 'new',
+    description: 'Start a new wordle game.',
+    options: [
+      {
+        name: 'id',
+        description: 'Wordle ID',
+        required: false,
+        type: "NUMBER",
+        minValue: 1,
+        maxValue: 100,
+      }
+    ]
+  });
 });
+
+client.on('interactionCreate', async (interaction) => {
+  if(!interaction.isCommand) return;
+
+  const { commandName, options } = interaction;
+
+  if(commandName === 'new'){
+    let id = options.getNumber('id');
+    wordle.newGame(id, interaction);
+  }
+})
 
 // --- When someone sends message --- \\
 client.on('messageCreate', msg =>{