+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 ### Changed
 - New game embed
 
+### Removed
+- Unescesairy intents
+
 ## [Release Alpha 1.2](https://github.com/Nevah5/wordlebot/releases/tag/a1.2) - 06.02.2022
 ### Added
 - Error message embed