+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Release Alpha 1.2](https://github.com/Nevah5/wordlebot/releases/tag/a1.2.1) - 06.02.2022
 ### Added
 - Help embed
+- Comments to `./src/modules/wordle.js`
 
 ### Changed
 - New game embed