+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog (Alpha Versions)
-## [Release Alpha 1.2](https://github.com/Nevah5/wordlebot/releases/tag/a1.2.1) - 06.02.2022
+## [Alpha 1.2.1](https://github.com/Nevah5/wordlebot/releases/tag/a1.2.1) - 06.02.2022
 ### Added
 - Help embed
 - Comments to `./src/modules/wordle.js`
+- User rich presence
 
 ### Changed
 - New game embed