+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog (Alpha Versions)
+## [Release Alpha 1.3](https://github.com/Nevah5/wordlebot/releases/tag/a1.2.1) - 06.02.2022
+### Added
+- Guild commands for testing
+
 ## [Alpha 1.2.1](https://github.com/Nevah5/wordlebot/releases/tag/a1.2.1) - 06.02.2022
 ### Added
 - Help embed